Show the added/removed keys in a git JSON diff
