+++ package.json
@@ -107,6 +107,8 @@
     "lodash.find": "^4.6.0",
     "lodash.includes": "^4.3.0",
     "lodash.isarray": "^4.0.0",
+    "lodash.isobject": "^2.4.1",
+    "lodash.keys": "^4.0.8",
     "node-fetch": "^1.6.3",
     "parse-diff": "^0.4.0",
     "rfc6902": "^1.3.0",
